add shared find_esv helper
