+++ src/Inversion_Workflow/Forward_Model/Calculate_Times_Bias.py
@@ -2,47 +2,7 @@ import numpy as np
 from numba import njit
 
 from geometry.ECEF_Geodetic import ECEF_Geodetic
-
-
-@njit
-def find_esv(beta, dz, dz_array, angle_array, esv_matrix):
-    """Interpolate effective sound velocities from a lookup table.
-
-    Parameters
-    ----------
-    beta : ndarray
-        Takeoff angles in degrees.
-    dz : ndarray
-        Vertical separation of receiver and source.
-    dz_array, angle_array, esv_matrix : ndarray
-        Discrete lookup grids defining the ESV table.
-
-    Returns
-    -------
-    ndarray
-        Interpolated effective sound velocities.
-    """
-    idx_closest_dz = np.empty_like(dz, dtype=np.int64)
-    idx_closest_beta = np.empty_like(beta, dtype=np.int64)
-
-    for i in range(len(dz)):
-        idx_closest_dz[i] = np.searchsorted(dz_array, dz[i], side="left")
-        if idx_closest_dz[i] < 0:
-            idx_closest_dz[i] = 0
-        elif idx_closest_dz[i] >= len(dz_array):
-            idx_closest_dz[i] = len(dz_array) - 1
-
-        idx_closest_beta[i] = np.searchsorted(angle_array, beta[i], side="left")
-        if idx_closest_beta[i] < 0:
-            idx_closest_beta[i] = 0
-        elif idx_closest_beta[i] >= len(angle_array):
-            idx_closest_beta[i] = len(angle_array) - 1
-
-    closest_esv = np.empty_like(dz, dtype=np.float64)
-    for i in range(len(dz)):
-        closest_esv[i] = esv_matrix[idx_closest_dz[i], idx_closest_beta[i]]
-
-    return closest_esv
+from Inversion_Workflow.Forward_Model.Find_ESV import find_esv
 
 
 @njit
